Refactor user-create-page test
