+++ test/interactors/AddressForm.js
@@ -1,0 +1,8 @@
+import { createInteractor, perform } from '@bigtest/interactor';
+
+export default createInteractor('list item')({
+  selector: 'li',
+  locators: {
+    findByHeaderLabel: element => element.querySelector('div[class^="addressLabel-"]').textContent,
+  }
+});
